Ported BMessage and dependencies

+++ headers/kernel/OS.h
@@ -27,7 +27,7 @@ extern "C" {
 /* System constants */
 
 #define B_OS_NAME_LENGTH	32
-#define B_PAGE_SIZE			getpagesize()
+#define B_PAGE_SIZE			((unsigned int) getpagesize())
 #define B_INFINITE_TIMEOUT	(9223372036854775807LL)
 
 enum {
